+++ internal/metarepos/storage_test.go
@@ -1,4 +1,4 @@
-package metadata_repository
+package metarepos
 
 import (
 	"bytes"