+++ tests/it/config.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 
-	"github.com/kakao/varlog/internal/metadata_repository"
+	"github.com/kakao/varlog/internal/metarepos"
 	"github.com/kakao/varlog/internal/varlogadm"
 	"github.com/kakao/varlog/pkg/types"
 	"github.com/kakao/varlog/pkg/util/testutil/ports"
@@ -25,12 +25,12 @@ const (
 	defaultStartVMS      = true
 )
 
-func defaultReportClientFactory() metadata_repository.ReporterClientFactory {
-	return metadata_repository.NewReporterClientFactory()
+func defaultReportClientFactory() metarepos.ReporterClientFactory {
+	return metarepos.NewReporterClientFactory()
 }
 
-func defaultStorageNodeManagementClientFactory() metadata_repository.StorageNodeManagementClientFactory {
-	return metadata_repository.NewStorageNodeManagementClientFactory()
+func defaultStorageNodeManagementClientFactory() metarepos.StorageNodeManagementClientFactory {
+	return metarepos.NewStorageNodeManagementClientFactory()
 }
 
 type config struct {
@@ -40,8 +40,8 @@ type config struct {
 	snapCount             int
 	collectorName         string
 	unsafeNoWAL           bool
-	reporterClientFac     metadata_repository.ReporterClientFactory
-	snManagementClientFac metadata_repository.StorageNodeManagementClientFactory
+	reporterClientFac     metarepos.ReporterClientFactory
+	snManagementClientFac metarepos.StorageNodeManagementClientFactory
 
 	numSN    int
 	numLS    int
@@ -131,13 +131,13 @@ func WithCollectorName(collector string) Option {
 	}
 }
 
-func WithReporterClientFactory(fac metadata_repository.ReporterClientFactory) Option {
+func WithReporterClientFactory(fac metarepos.ReporterClientFactory) Option {
 	return func(c *config) {
 		c.reporterClientFac = fac
 	}
 }
 
-func WithStorageNodeManagementClientFactory(fac metadata_repository.StorageNodeManagementClientFactory) Option {
+func WithStorageNodeManagementClientFactory(fac metarepos.StorageNodeManagementClientFactory) Option {
 	return func(c *config) {
 		c.snManagementClientFac = fac
 	}